# docs-standalone/openapi.yaml
openapi: 3.0.3
info:
  title: Examify API Documentation
  version: 1.0.0
  description: >-
    **Examify** is a production-grade, self-hosted AI Exam Generation &
    Assessment Platform built on Next.js 15 (App Router), MariaDB (Prisma ORM),
    Redis, and multi-model AI adapters (Google Gemini, OpenAI, Anthropic Claude,
    and Self-Hosted/vLLM/Ollama).


    ### Core Security & Platform Highlights:

    - **Zero Cloudflare Dependencies**: Self-hosted Redis Sliding Window rate
    limiting and MaxMind GeoIP resolution.

    - **Strict Role-Based Access Control**: `ADMIN` and `TEACHER` roles with
    token blacklisting in Redis.

    - **Two-Factor Authentication (2FA)**: Built-in TOTP (Google Authenticator)
    with QR code setup and challenge verification.

    - **Anti-Prompt Injection Guard**: Advanced multi-tier sanitization and
    jailbreak defense for teacher custom instructions.

    - **Permanent Initialization Lock**: One-time initial setup endpoint locked
    permanently once root admin is created.

    - **Comprehensive Audit Trail**: Immutable system logs recording security
    events, authentication attempts, AI synthesis, and administrative updates.
  contact:
    name: Examify Engineering Team
    email: admin@examify.local
    url: http://localhost:3000
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Examify API v1 Base Path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Provide your JWT bearer token in the Authorization header (e.g.
        `Bearer eyJhbGciOi...`).
  schemas:
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
        data:
          type: object
    StandardError:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found or validation failure.
        code:
          type: string
          example: VALIDATION_ERROR
        errors:
          type: object
          description: Field-level validation error details
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d3b07384-d113-460d-9b16-f288e2c4908a
        email:
          type: string
          format: email
          example: teacher@examify.local
        fullName:
          type: string
          example: ThS. Nguyễn Văn An
        role:
          type: string
          enum:
            - ADMIN
            - TEACHER
          example: TEACHER
        school:
          type: string
          example: THPT Chuyên Trần Phú
        twoFactorEnabled:
          type: boolean
          example: false
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    CognitiveMatrix:
      type: object
      description: Bloom taxonomy percentage distribution (must sum to 100%).
      properties:
        knowledge:
          type: number
          minimum: 0
          maximum: 100
          example: 40
          description: Nhận biết (%)
        comprehension:
          type: number
          minimum: 0
          maximum: 100
          example: 30
          description: Thông hiểu (%)
        application:
          type: number
          minimum: 0
          maximum: 100
          example: 20
          description: Vận dụng (%)
        highApplication:
          type: number
          minimum: 0
          maximum: 100
          example: 10
          description: Vận dụng cao (%)
    QuestionOption:
      type: object
      properties:
        key:
          type: string
          example: A
        text:
          type: string
          example: x = 2 hoặc x = -3
        isCorrect:
          type: boolean
          example: false
    Question:
      type: object
      properties:
        id:
          type: string
          example: q-1
        questionNumber:
          type: integer
          example: 1
        type:
          type: string
          enum:
            - multiple_choice
            - true_false
            - short_answer
            - essay
          example: multiple_choice
        cognitiveLevel:
          type: string
          enum:
            - KNOWLEDGE
            - COMPREHENSION
            - APPLICATION
            - HIGH_APPLICATION
          example: COMPREHENSION
        level:
          type: string
          example: Thông hiểu
        content:
          type: string
          example: "Cho hàm số \\(y = f(x)\\) có bảng biến thiên như hình vẽ. Số điểm cực
            trị của hàm số là:"
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correctAnswer:
          type: string
          example: B
        explanation:
          type: string
          example: Đạo hàm đổi dấu 2 lần qua các điểm cực trị.
        points:
          type: number
          example: 0.25
    ExamItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b1b52817-820d-41d9-adfe-55c8293546c1
        title:
          type: string
          example: Đề thi Khảo sát chất lượng Toán 12 - Học kỳ 1
        subject:
          type: string
          example: Toán học
        grade:
          type: string
          example: Lớp 12
        term:
          type: string
          example: Học kỳ 1
        duration:
          type: integer
          example: 90
        status:
          type: string
          enum:
            - approved
            - review
            - draft
          example: approved
        totalQuestions:
          type: integer
          example: 22
        matrix:
          $ref: "#/components/schemas/CognitiveMatrix"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuditLogItem:
      type: object
      properties:
        id:
          type: string
          example: log-102938
        action:
          type: string
          example: EXAM_GENERATED
        entity:
          type: string
          example: Exam
        entityId:
          type: string
          example: b1b52817-820d-41d9-adfe-55c8293546c1
        userId:
          type: string
          example: usr-admin-01
        user:
          type: object
          properties:
            fullName:
              type: string
              example: Nguyễn Văn An
            email:
              type: string
              example: an.nguyen@spt.edu.vn
        clientIp:
          type: string
          example: 14.161.22.45
        country:
          type: string
          example: VN
        userAgent:
          type: string
          example: Mozilla/5.0 (Windows NT 10.0; Win64; x64)
        details:
          type: object
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication & Profile
      summary: Đăng nhập người dùng & Cấp phiên JWT
      description: "Xác thực tài khoản bằng email và mật khẩu. Nếu tài khoản bật xác
        thực hai yếu tố (2FA), hệ thống trả về mã `requires2FA: true` và
        `tempToken` để tiếp tục xác thực TOTP."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: teacher@examify.local
                password:
                  type: string
                  example: TeacherP@ss123
      responses:
        "200":
          description: Đăng nhập thành công (hoặc yêu cầu mã 2FA)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1Ni...
                      user:
                        $ref: "#/components/schemas/UserResponse"
                      requires2FA:
                        type: boolean
                        example: false
                      tempToken:
                        type: string
                        description: Token tạm thời dùng cho bước verify 2FA
        "401":
          description: Email hoặc mật khẩu không chính xác
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
  /exams/generate:
    post:
      tags:
        - Exams & Synthesis
      summary: Khởi tạo đề thi tự động bằng Trí tuệ Nhân tạo (AI)
      description: Sinh trọn vẹn một đề thi chuẩn Bộ GD&ĐT dựa trên phân phối ma trận
        nhận thức Bloom, cấp học, môn thi, và hướng dẫn tùy biến của giáo viên
        (có bộ lọc PromptGuard chống Jailbreak).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - curriculum
                - examType
              properties:
                title:
                  type: string
                  minLength: 3
                  example: Đề thi Khảo sát Toán 12
                curriculum:
                  type: object
                  required:
                    - level
                    - grade
                    - subject
                    - semester
                  properties:
                    level:
                      type: string
                      example: THPT
                    grade:
                      type: string
                      example: Lớp 12
                    subject:
                      type: string
                      example: Toán học
                    semester:
                      type: string
                      example: Học kỳ 1
                    topic:
                      type: string
                      example: Ứng dụng đạo hàm khảo sát hàm số
                    lesson:
                      type: string
                      example: Cực trị của hàm số
                examType:
                  type: string
                  enum:
                    - MIN_15
                    - MID_TERM
                    - FINAL_TERM
                    - ENTRANCE
                  example: MID_TERM
                durationMinutes:
                  type: integer
                  default: 45
                  example: 45
                totalPoints:
                  type: number
                  default: 10
                  example: 10
                totalQuestions:
                  type: integer
                  default: 20
                  example: 20
                cognitiveMatrix:
                  $ref: "#/components/schemas/CognitiveMatrix"
                customPrompt:
                  type: string
                  maxLength: 1500
                  example: Tăng cường các câu hỏi liên hệ thực tiễn và tối ưu hóa cho kỳ thi THPT
                    Quốc gia.
                aiProvider:
                  type: string
                  enum:
                    - gemini
                    - openai
                    - claude
                    - self-hosted
                    - custom-mcp
                  example: gemini
                aiModel:
                  type: string
                  example: gemini-2.0-flash
                language:
                  type: string
                  default: Vietnamese
      responses:
        "201":
          description: Tạo đề thi thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Tạo đề thi thành công
                  data:
                    type: object
                    properties:
                      examId:
                        type: string
                        format: uuid
                      title:
                        type: string
                      status:
                        type: string
                        example: COMPLETED
                      examData:
                        $ref: "#/components/schemas/ExamItem"
        "400":
          description: Phát hiện mã độc Prompt Injection hoặc tham số không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "429":
          description: Vượt quá giới hạn gọi AI đồng thời hoặc giới hạn lượt yêu cầu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "502":
          description: Lỗi phản hồi từ nhà cung cấp mô hình AI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
  /ai/settings:
    get:
      tags:
        - AI Configuration
      summary: Xem cấu hình AI hệ thống hiện tại (ADMIN only)
      description: Lấy thông tin nhà cung cấp mặc định, model đang dùng, số lượng API
        keys đã nạp và trạng thái trạm tự lưu trữ.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cấu hình AI hiện hành
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      defaultProvider:
                        type: string
                        example: gemini
                      geminiModel:
                        type: string
                        example: gemini-2.0-flash
                      geminiKeysCount:
                        type: integer
                        example: 2
                      openaiModel:
                        type: string
                        example: gpt-4o-mini
                      openaiKeysCount:
                        type: integer
                        example: 1
                      claudeModel:
                        type: string
                        example: claude-3-5-sonnet
                      claudeKeysCount:
                        type: integer
                        example: 0
                      selfHostedBaseUrl:
                        type: string
                        example: http://localhost:8000/v1
                      dailyTokenLimit:
                        type: integer
                        example: 500000
    post:
      tags:
        - AI Configuration
      summary: Cập nhật cấu hình & API Keys của các mô hình AI (ADMIN only)
      description: Cập nhật nhà cung cấp mặc định, nạp danh sách xoay vòng API Keys
        (Multi-Key Pool Rotation) và điều chỉnh tham số nhiệt độ (temperature).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                defaultProvider:
                  type: string
                  enum:
                    - gemini
                    - openai
                    - claude
                    - self-hosted
                geminiApiKeys:
                  type: string
                  description: Các key phân tách bằng dấu phẩy hoặc xuống dòng
                geminiModel:
                  type: string
                  example: gemini-2.0-flash
                openaiApiKeys:
                  type: string
                  description: Các key phân tách bằng dấu phẩy
                openaiModel:
                  type: string
                  example: gpt-4o-mini
                claudeApiKeys:
                  type: string
                claudeModel:
                  type: string
                  example: claude-3-5-sonnet
                selfHostedBaseUrl:
                  type: string
                  example: http://localhost:11434/v1
                temperature:
                  type: number
                  minimum: 0
                  maximum: 2
                  default: 0.7
      responses:
        "200":
          description: Lưu cấu hình AI thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
...
  /admin/logs:
    get:
      tags:
        - Admin - Audit Logs
      summary: Truy vấn nhật ký kiểm toán hệ thống (Audit Logs) (ADMIN only)
      description: Lấy danh sách nhật ký kiểm toán bất biến (Immutable Audit Logs) ghi
        nhận các sự kiện bảo mật, đăng nhập, sinh đề thi, và thay đổi cài đặt.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: action
          in: query
          required: false
          schema:
            type: string
            example: EXAM_GENERATED
      responses:
        "200":
          description: Danh sách nhật ký kiểm toán
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogItem"
    post:
      tags:
        - Admin - Audit Logs
      summary: Tìm kiếm & Lọc nhật ký kiểm toán nâng cao (ADMIN only)
      description: Tìm kiếm nhật ký kiểm toán theo dải thời gian, IP người dùng, loại
        sự kiện hoặc từ khóa chi tiết.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                userId:
                  type: string
                clientIp:
                  type: string
                keyword:
                  type: string
                limit:
                  type: integer
                  default: 100
      responses:
        "200":
          description: Kết quả lọc nhật ký
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogItem"
...