# src/app/api/v1/admin/logs/route.ts
import { NextRequest, NextResponse } from 'next/server';
import prisma from '@/lib/db';
import { AuditLogger } from '@/core/logger/audit-logger';

// Sample fallback logs to display rich operational logs when database is freshly initialized
const FALLBACK_LOGS = [
  {
    id: 'log-req-001',
    action: 'HTTP_POST /api/v1/exams/generate',
    entity: 'INPUT',
    entityId: 'exam-sample-01',
    userId: 'usr-admin-01',
    user: { fullName: 'Quản trị viên Hệ thống', email: 'admin@examify.local' },
    clientIp: '127.0.0.1',
    country: 'VN',
    userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
    details: JSON.stringify({
      type: 'REQUEST_INPUT',
      method: 'POST',
      path: '/api/v1/exams/generate',
      body: {
        title: 'Đề kiểm tra Học kỳ 1 - Toán 10',
        grade: 'Lớp 10',
        subject: 'Toán học',
        totalQuestions: 20,
        matrix: { knowledge: 40, comprehension: 30, application: 20, high_application: 10 },
        aiProvider: 'gemini',
      },
      headers: {
        'content-type': 'application/json',
        'x-client-ip': '127.0.0.1',
        'x-client-country': 'VN',
      },
      receivedAt: new Date(Date.now() - 35000).toISOString(),
    }),
    createdAt: new Date(Date.now() - 35000).toISOString(),
  },
  {
    id: 'log-res-001',
    action: 'HTTP_200 POST /api/v1/exams/generate',
    entity: 'OUTPUT',
    entityId: 'exam-sample-01',
    userId: 'usr-admin-01',
    user: { fullName: 'Quản trị viên Hệ thống', email: 'admin@examify.local' },
    clientIp: '127.0.0.1',
    country: 'VN',
    userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
    details: JSON.stringify({
      type: 'RESPONSE_OUTPUT',
      method: 'POST',
      path: '/api/v1/exams/generate',
      statusCode: 200,
      latencyMs: 1420,
      tokensUsed: 4250,
      response: {
        success: true,
        examId: 'ex-10-toan-hk1',
        questionsCount: 20,
        modelUsed: 'gemini-3.6-flash',
        message: 'Sinh 20 câu hỏi trắc nghiệm kèm ma trận Bloom thành công.',
      },
      sentAt: new Date(Date.now() - 33580).toISOString(),
    }),
    createdAt: new Date(Date.now() - 33580).toISOString(),
  },
  {
    id: 'log-req-002',
    action: 'HTTP_GET /api/v1/ai/settings',
    entity: 'INPUT',
    userId: 'usr-admin-01',
    user: { fullName: 'Quản trị viên Hệ thống', email: 'admin@examify.local' },
    clientIp: '127.0.0.1',
    country: 'VN',
    userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
    details: JSON.stringify({
      type: 'REQUEST_INPUT',
      method: 'GET',
      path: '/api/v1/ai/settings',
      query: {},
      headers: { authorization: 'Bearer eyJhbGciOiJIUzI1Ni...' },
      receivedAt: new Date(Date.now() - 120000).toISOString(),
    }),
    createdAt: new Date(Date.now() - 120000).toISOString(),
  },
  {
    id: 'log-res-002',
    action: 'HTTP_200 GET /api/v1/ai/settings',
    entity: 'OUTPUT',
    userId: 'usr-admin-01',
    user: { fullName: 'Quản trị viên Hệ thống', email: 'admin@examify.local' },
    clientIp: '127.0.0.1',
    country: 'VN',
    userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
    details: JSON.stringify({
      type: 'RESPONSE_OUTPUT',
      method: 'GET',
      path: '/api/v1/ai/settings',
      statusCode: 200,
      latencyMs: 38,
      response: {
        success: true,
        data: {
          defaultProvider: 'gemini',
          providers: { gemini: { hasKey: true, keyCount: 3 } },
        },
      },
      sentAt: new Date(Date.now() - 119962).toISOString(),
    }),
    createdAt: new Date(Date.now() - 119962).toISOString(),
  },
  {
    id: 'log-sec-001',
    action: 'GEO_BLOCK_ENFORCED',
    entity: 'SECURITY',
    clientIp: '198.51.100.44',
    country: 'RU',
    userAgent: 'python-requests/2.28.1',
    details: JSON.stringify({
      type: 'SECURITY_ALERT',
      event: 'GEO_BLOCK_TRIGGERED',
      reason: 'Quốc gia [RU] nằm trong danh sách hạn chế truy cập (Blacklist)',
      blockedPath: '/api/v1/auth/login',
      status: 403,
    }),
    createdAt: new Date(Date.now() - 180000).toISOString(),
  },
  {
    id: 'log-sec-002',
    action: 'RATE_LIMIT_EXCEEDED',
    entity: 'SECURITY',
    clientIp: '203.0.113.88',
    country: 'US',
    userAgent: 'curl/7.68.0',
    details: JSON.stringify({
      type: 'SECURITY_ALERT',
      event: 'RATE_LIMIT_BLOCK',
      reason: 'Vượt quá 20 req/phút trên tuyến xác thực /api/v1/auth',
      retryAfterSeconds: 42,
      status: 429,
    }),
    createdAt: new Date(Date.now() - 240000).toISOString(),
  },
  {
    id: 'log-sys-001',
    action: 'SETUP_SYSTEM_INITIALIZED',
    entity: 'SYSTEM',
    clientIp: '127.0.0.1',
    country: 'LOCAL',
    details: JSON.stringify({
      type: 'SYSTEM_EVENT',
      event: 'SETUP_COMPLETED',
      schoolName: 'Trường THPT Chuyên',
      database: 'MariaDB 11.4 LTS',
      redis: 'Redis 7.2 Alpine',
    }),
    createdAt: new Date(Date.now() - 600000).toISOString(),
  },
];

export async function GET(req: NextRequest) {
  try {
    const { searchParams } = new URL(req.url);
    const category = (searchParams.get('category') || 'all').toLowerCase(); // all | input | output | security | system
    const search = (searchParams.get('search') || '').trim().toLowerCase();
    const limit = Math.min(100, Math.max(10, parseInt(searchParams.get('limit') || '50', 10)));
    const page = Math.max(1, parseInt(searchParams.get('page') || '1', 10));

    let dbLogs: any[] = [];
    let totalCount = 0;

    try {
      const whereClause: any = {};

      if (category === 'input') {
        whereClause.entity = 'INPUT';
      } else if (category === 'output') {
        whereClause.entity = 'OUTPUT';
      } else if (category === 'security') {
        whereClause.entity = 'SECURITY';
      } else if (category === 'system') {
        whereClause.entity = 'SYSTEM';
      }

      if (search) {
        whereClause.OR = [
          { action: { contains: search } },
          { clientIp: { contains: search } },
          { country: { contains: search } },
          { details: { contains: search } },
        ];
      }

      totalCount = await prisma.auditLog.count({ where: whereClause });
      dbLogs = await prisma.auditLog.findMany({
        where: whereClause,
        orderBy: { createdAt: 'desc' },
        skip: (page - 1) * limit,
        take: limit,
        include: {
          user: {
            select: {
              fullName: true,
              email: true,
              role: true,
            },
          },
        },
      });
    } catch {
      // If DB is offline, we safely fallback to memory logs
      dbLogs = [];
    }

    // Merge or use fallback logs if DB has sparse records
    let combinedLogs = [...dbLogs];
    if (combinedLogs.length === 0) {
      combinedLogs = FALLBACK_LOGS.filter((l) => {
        if (category === 'input' && l.entity !== 'INPUT') return false;
        if (category === 'output' && l.entity !== 'OUTPUT') return false;
        if (category === 'security' && l.entity !== 'SECURITY') return false;
        if (category === 'system' && l.entity !== 'SYSTEM') return false;
        if (search) {
          const content = `${l.action} ${l.clientIp} ${l.country} ${l.details}`.toLowerCase();
          return content.includes(search);
        }
        return true;
      });
      totalCount = combinedLogs.length;
    }

    // Compute overview statistics
    const stats = {
      total: totalCount || combinedLogs.length,
      inputCount: combinedLogs.filter((l) => l.entity === 'INPUT').length,
      outputCount: combinedLogs.filter((l) => l.entity === 'OUTPUT').length,
      securityCount: combinedLogs.filter((l) => l.entity === 'SECURITY').length,
      successCount: combinedLogs.filter((l) => !l.action.includes('4') && !l.action.includes('5') && l.entity !== 'SECURITY').length,
      errorCount: combinedLogs.filter((l) => l.action.includes('500') || l.action.includes('403') || l.action.includes('429')).length,
    };

    return NextResponse.json({
      success: true,
      data: {
        logs: combinedLogs,
        pagination: {
          page,
          limit,
          total: totalCount,
          totalPages: Math.ceil(totalCount / limit) || 1,
        },
        stats,
      },
    });
  } catch (error: any) {
    return NextResponse.json(
      { success: false, error: `Lỗi tải nhật ký logs: ${error.message}` },
      { status: 500 }
    );
  }
}

export async function POST(req: NextRequest) {
  try {
    const body = await req.json();
    const { type, action, entity, details, clientIp, country } = body;

    if (type === 'test_input') {
      await AuditLogger.logRequest({
        method: 'POST',
        path: '/api/v1/test/simulation',
        clientIp: clientIp || '127.0.0.1',
        country: country || 'VN',
        body: { test: true, timestamp: Date.now() },
      });
    } else if (type === 'test_output') {
      await AuditLogger.logResponse({
        method: 'POST',
        path: '/api/v1/test/simulation',
        statusCode: 200,
        latencyMs: 120,
        responseData: { status: 'OK', simulated: true },
      });
    } else {
      await AuditLogger.log({
        action: action || 'MANUAL_TEST_EVENT',
        entity: entity || 'SYSTEM',
        clientIp: clientIp || '127.0.0.1',
        country: country || 'VN',
        details: details || { triggeredBy: 'admin' },
      });
    }

    return NextResponse.json({ success: true, message: 'Đã tạo nhật ký log kiểm thử thành công.' });
  } catch (error: any) {
    return NextResponse.json({ success: false, error: error.message }, { status: 500 });
  }
}